requirements.txt
alembic==1.5.8
Flask==1.1.2
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
form==0.0.1
Jinja2==2.11.2
python-telegram-bot==13.4.1
requests==2.25.1
SQLAlchemy==1.3.23
SQLAlchemy-serializer==1.3.4.4
sqlparse==0.4.1
Werkzeug==1.0.1
WTForms==2.3.3
